+++ src/lib/types.ts
@@ -20,8 +20,12 @@ export type TorrentSourceId =
   | 'yourbittorrent'
   | 'torrentscsv'
   | 'eztv'
+  | 'yts'
+  | 'therarbg'
+  | 'torrentdownloads'
   | 'animetosho'
   | 'nyaa'
+  | 'tokyotosho'
   | 'limetorrents'
   | 'jackett';
 
@@ -31,8 +35,12 @@ export interface TorrentSources {
   yourbittorrent: boolean;
   torrentscsv: boolean;
   eztv: boolean;
+  yts: boolean;
+  therarbg: boolean;
+  torrentdownloads: boolean;
   animetosho: boolean;
   nyaa: boolean;
+  tokyotosho: boolean;
   limetorrents: boolean;
   jackett: boolean;
 }
@@ -397,8 +405,12 @@ export const DEFAULT_TORRENT_SOURCES: TorrentSources = {
   yourbittorrent: true,
   torrentscsv: true,
   eztv: true,
-  animetosho: true,
+  yts: true,
+  therarbg: true,
+  torrentdownloads: true,
+  animetosho: false,
   nyaa: true,
+  tokyotosho: true,
   limetorrents: true,
   jackett: false,
 };